refactor(keyring-api): flatten folders
