--- packages/keyring-api/src/api/index.ts
@@ -1,7 +1,0 @@
-export * from './account';
-export * from './balance';
-export * from './caip';
-export * from './export';
-export type * from './keyring';
-export * from './request';
-export * from './response';
