+++ packages/keyring-api/src/index.ts
@@ -1,2 +1,8 @@
-export * from './api';
+export * from './account';
+export * from './balance';
+export * from './caip';
+export * from './export';
+export type * from './keyring';
+export * from './request';
+export * from './response';
 export type * from './contexts';
