+++ packages/keyring-api/src/keyring.ts
